+++ evo_prot_grad/common/embeddings.py
@@ -35,7 +35,6 @@ class OneHotEmbedding(nn.Module):
         nn_embeddings: nn.Embedding
     ):
         super().__init__()
-        print("Using local OneHotEmbedding class")
         self.weight = nn_embeddings.weight
         self.one_hots = None
          
@@ -49,20 +48,15 @@ class OneHotEmbedding(nn.Module):
         Returns:
             embeddings (torch.FloatTensor): Amino acid embeddings of shape [batch_size, max_sequence_len, embedding_dim].
         """
+        weights_dtype = self.weight.dtype  # could be float16 if using mixed precision
+        high_precision = torch.float32  # optionally float64 ??
         # convert input_ids to one_hots
         # one_hots is a torch.FloatTensor of shape [batch_size, max_sequence_len, vocab_size]
-        #one_hots = torch.nn.functional.one_hot(input_ids, num_classes=self.weight.shape[0])
-        #one_hots = one_hots.to(dtype=self.weight.dtype) # to be compatible with half precision ESM2
-        # cache the one_hots
-        #self.one_hots = one_hots.float().requires_grad_()
-        #self.one_hots = one_hots.requires_grad_()
-        #return self.one_hots @ self.weight
         one_hots = torch.nn.functional.one_hot(input_ids, num_classes=self.weight.shape[0])
-        one_hots = one_hots.to(dtype=torch.float32)  # Ensure one_hots are in float32 for gradient computation
+        one_hots = one_hots.to(dtype=high_precision)  # Ensure one_hots are in float32 for gradient computation
         # Cache the one_hots
         self.one_hots = one_hots.requires_grad_()
-        # Compute the embeddings and convert back to float16 if necessary
-        embeddings = self.one_hots @ self.weight.to(dtype=torch.float32)
-        if self.weight.dtype == torch.float16:
-            embeddings = embeddings.to(dtype=torch.float16)
+        # Compute the embeddings and convert back to low precision if necessary
+        embeddings = self.one_hots @ self.weight.to(dtype=high_precision)
+        embeddings = embeddings.to(dtype=weights_dtype)
         return embeddings